refactor: move dialogs into a separate folder ♻️
